# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Bid Change Validator API
  version: 0.1.0
paths:
...
  /api/v1/notices/sync:
    post:
      tags:
      - bid notices
      summary: Sync Notices
      operationId: sync_notices_api_v1_notices_sync_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoticeSyncRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoticeCollectionRunRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/notices/collection-runs:
    get:
      tags:
      - bid notices
      summary: List Collection Runs
      operationId: list_collection_runs_api_v1_notices_collection_runs_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NoticeCollectionRunRead'
                title: Response List Collection Runs Api V1 Notices Collection Runs
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/notices:
    get:
      tags:
      - bid notices
      summary: Search Notices
      operationId: search_notices_api_v1_notices_get
      parameters:
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Q
      - name: business_type
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/BusinessType'
          - type: 'null'
          title: Business Type
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BidNoticeSearchResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/notices/documents/extract-pending:
    post:
      tags:
      - bid notices
      summary: Extract Pending Notice Documents
      operationId: extract_pending_notice_documents_api_v1_notices_documents_extract_pending_post
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 100
          title: Limit
      - name: retry_failed
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Retry Failed
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentExtractionBatchRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/notices/{notice_id}/versions/{version_number}/documents/{document_id}/text:
    get:
      tags:
      - bid notices
      summary: Get Notice Document Text
      operationId: get_notice_document_text_api_v1_notices__notice_id__versions__version_number__documents__document_id__text_get
      parameters:
      - name: notice_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Notice Id
      - name: version_number
        in: path
        required: true
        schema:
          type: integer
          title: Version Number
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Document Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoticeDocumentTextRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/notices/{notice_id}/versions/{version_number}/documents/{document_id}/source:
    get:
      tags:
      - bid notices
      summary: Get Notice Document Render Source
      operationId: get_notice_document_render_source_api_v1_notices__notice_id__versions__version_number__documents__document_id__source_get
      parameters:
      - name: notice_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Notice Id
      - name: version_number
        in: path
        required: true
        schema:
          type: integer
          title: Version Number
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Document Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/notices/{notice_id}/versions/{version_number}/documents/{document_id}/content:
    get:
      tags:
      - bid notices
      summary: Download Notice Document
      operationId: download_notice_document_api_v1_notices__notice_id__versions__version_number__documents__document_id__content_get
      parameters:
      - name: notice_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Notice Id
      - name: version_number
        in: path
        required: true
        schema:
          type: integer
          title: Version Number
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Document Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/notices/{notice_id}/versions/{version_number}/documents/{document_id}/preview:
    get:
      tags:
      - bid notices
      summary: Preview Notice Document
      operationId: preview_notice_document_api_v1_notices__notice_id__versions__version_number__documents__document_id__preview_get
      parameters:
      - name: notice_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Notice Id
      - name: version_number
        in: path
        required: true
        schema:
          type: integer
          title: Version Number
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Document Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/notices/{notice_id}:
    get:
      tags:
      - bid notices
      summary: Get Notice
      operationId: get_notice_api_v1_notices__notice_id__get
      parameters:
      - name: notice_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Notice Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BidNoticeDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/notices/{notice_id}/versions:
    get:
      tags:
      - bid notices
      summary: List Notice Versions
      operationId: list_notice_versions_api_v1_notices__notice_id__versions_get
      parameters:
      - name: notice_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Notice Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BidNoticeVersionRead'
                title: Response List Notice Versions Api V1 Notices  Notice Id  Versions
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    BusinessType:
      type: string
      enum:
      - SERVICE
      - GOODS
      - CONSTRUCTION
      - FOREIGN
      - OTHER
      title: BusinessType
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    NoticeCollectionRunRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        business_type:
          $ref: '#/components/schemas/BusinessType'
        inquiry_type:
          $ref: '#/components/schemas/NoticeInquiryType'
        window_started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Window Started At
        window_ended_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Window Ended At
        status:
          type: string
          title: Status
        api_calls:
          type: integer
          title: Api Calls
        fetched_count:
          type: integer
          title: Fetched Count
        created_count:
          type: integer
          title: Created Count
        new_version_count:
          type: integer
          title: New Version Count
        unchanged_count:
          type: integer
          title: Unchanged Count
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        started_at:
          type: string
          format: date-time
          title: Started At
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
      type: object
      required:
      - id
      - business_type
      - inquiry_type
      - window_started_at
      - window_ended_at
      - status
      - api_calls
      - fetched_count
      - created_count
      - new_version_count
      - unchanged_count
      - error_message
      - started_at
      - completed_at
      title: NoticeCollectionRunRead
...
    NoticeInquiryType:
      type: string
      enum:
      - REGISTERED
      - CHANGED
      - NOTICE_NUMBER
      title: NoticeInquiryType
    NoticeSyncRequest:
      properties:
        business_type:
          $ref: '#/components/schemas/BusinessType'
        inquiry_type:
          $ref: '#/components/schemas/NoticeInquiryType'
          default: REGISTERED
        window_started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Window Started At
        window_ended_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Window Ended At
        bid_notice_no:
          anyOf:
          - type: string
          - type: 'null'
          title: Bid Notice No
        page_size:
          type: integer
          maximum: 999.0
          minimum: 1.0
          title: Page Size
          default: 100
        max_pages:
          type: integer
          maximum: 100.0
          minimum: 1.0
          title: Max Pages
          default: 10
      type: object
      required:
      - business_type
      title: NoticeSyncRequest
...